fill rates is optional

+++ scripts/helper_scripts/IRF.optimizeTMVAforGammaHadronSeparation_sub.sh
@@ -25,10 +25,12 @@ mkdir -p $TEMPDIR
 ls -1 ${PREDIR}/${CUT}/*.anasum.root > ${TEMPDIR}/anasum.list
 
 OBSTIME="5."
-if [[ $CUT == *"Moderate"* ]] || [[ $CUT == *"Hard"* ]]; then
+if [[ $CUT == *"Moderate"* ]]; then
     OBSTIME="5."
 elif [[ $CUT == *"Soft"* ]]; then
     OBSTIME="1."
+elif [[ $CUT == *"Hard"* ]]; then
+    OBSTIME="5."
 fi
 
 # effective area - fill path
@@ -44,17 +46,21 @@ mkdir -p ${WDIR}
 # rates files
 RATEFILE="${WDIR}/rates_${EPAT}"
 
-rm -f ${RATEFILE}.log
+CALCULATERATEFILES="TRUE"
+if [[ $CALCULATERATEFILES == "TRUE" ]];
+then
+    rm -f ${RATEFILE}.log
 
-# calculate rates from Crab Nebula and from background rates
-rm -f ${MVADIR}/rates.log
-"$EVNDISPSYS"/bin/calculateCrabRateFromMC \
-    ${EFFAREA} \
-    ${RATEFILE}.root \
-    ${DEADTIME} \
-    ${VERITAS_EVNDISP_AUX_DIR}/ParameterFiles/TMVA.BDT.runparameter \
-    ${TEMPDIR}/anasum.list \
-    > ${RATEFILE}.log
+    # calculate rates from Crab Nebula and from background rates
+    rm -f ${MVADIR}/rates.log
+    "$EVNDISPSYS"/bin/calculateCrabRateFromMC \
+        ${EFFAREA} \
+        ${RATEFILE}.root \
+        ${DEADTIME} \
+        ${VERITAS_EVNDISP_AUX_DIR}/ParameterFiles/TMVA.BDT.runparameter \
+        ${TEMPDIR}/anasum.list \
+        > ${RATEFILE}.log
+fi
 
 # optimize cuts
 echo "optimize cuts..."
